ETag depends on vary uri
